Update page layout using flex
Use flex on body, set direction to column. Add
flex to header, content, and footer sections. Use
margin-auto to make the header and footer stay
at the top and bottom. Also, use flex to make
the left and right links stay on the sides of the
page when resizing. All of these changes are to
match the final outcome per TOP instructions for
this exercise.